+++ addons/web/views/webclient_templates.xml
@@ -489,6 +489,15 @@
             }
         </style>
 
+        <script type="text/javascript" src="/web/static/tests/helpers/test_utils_create.js"></script>
+        <script type="text/javascript" src="/web/static/tests/helpers/test_utils_dom.js"></script>
+        <script type="text/javascript" src="/web/static/tests/helpers/test_utils_fields.js"></script>
+        <script type="text/javascript" src="/web/static/tests/helpers/test_utils_form.js"></script>
+        <script type="text/javascript" src="/web/static/tests/helpers/test_utils_graph.js"></script>
+        <script type="text/javascript" src="/web/static/tests/helpers/test_utils_kanban.js"></script>
+        <script type="text/javascript" src="/web/static/tests/helpers/test_utils_mock.js"></script>
+        <script type="text/javascript" src="/web/static/tests/helpers/test_utils_modal.js"></script>
+        <script type="text/javascript" src="/web/static/tests/helpers/test_utils_pivot.js"></script>
         <script type="text/javascript" src="/web/static/tests/helpers/test_utils.js"></script>
         <script type="text/javascript" src="/web/static/tests/helpers/mock_server.js"></script>
